fix(auth, android): fixed user collision handling with apple sign-in (#4487)
This fixes an issue with Apple sign on Android. iOS was fixed with this commit. iOS was fixed with this PR https://github.com/invertase/react-native-firebase/pull/4359. Note the solution is different. On iOS, it looks like we bubble the auth credentials back to js land and then retry login. Here, we attempt another login on auth collision errors