build(e2e, ios): integrate xcbeautify, with correct compile failure status report
the current xcode build is still very noisy

previously there was an issue with combination of xcpretty and CI, such that
compile failures were not detected and failing CI.

xcpretty is abandoned but xcbeautify lives. The built-in `-quiet` is too quiet.

I believe this combo of set -o pipefail and xcbeautify should do it.
It passes local negative- and positive-testing on build status reporting